Do not delete tracking dismissed flag when tracking is on

+++ installer.php
@@ -15,7 +15,8 @@
         if ( $this->installed_version != self::DB_VERSION ) {
             $this->_database_schema();
 
-            delete_option( 'wpbdp-tracking-dismissed' );
+            if ( !wpbdp_get_option( 'tracking-on', false ) )
+                delete_option( 'wpbdp-tracking-dismissed' );
         }
 
         if ( $this->installed_version ) {
